pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
    "asyncpg>=0.30.0",
    "anthropic>=0.40.0",
    "pillow>=11.0.0",
    "watchdog>=6.0.0",
    "python-magic>=0.4.27",
    "piexif>=1.1.3",
    "aioimaplib>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "transformers>=4.47.0",
    "torch>=2.5.0",
    "fastmcp>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
